backend/requirements.txt
asgiref==3.3.1
astroid==2.5.2
certifi==2020.12.5
chardet==4.0.0
currency-symbols==2.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-redis==4.12.1
djangorestframework==3.12.4
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.19.5
pandas==1.1.5
pylint==2.7.4
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
crontab==0.22.9
gunicorn==20.1.0
whitenoise==5.2.0